-feat: 1、完成转换器的提取BaseConverter；2、完成实体类数字类型的Integer转换为Long；3、Jwt鉴权完善，且自动更新；4、完成用户上下文创建的全流程，创建VO，完成转换器，在获取用户授权时如果发现未存在于数据库时，自动创建数据库记录；5、完成域名路径拦截器（对特定路径进行拦截鉴权）

+++ src/main/java/com/shiroko/repository/entity/CourseRecord.java
@@ -22,7 +22,7 @@ public class CourseRecord implements Serializable {
      * 表id
      */
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 学生姓名
@@ -40,13 +40,13 @@ public class CourseRecord implements Serializable {
      * 课时总数
      */
     @TableField(value = "course_total_time")
-    private Integer courseTotalTime;
+    private Long courseTotalTime;
 
     /**
      * 课程剩余次数
      */
     @TableField(value = "course_rest_time")
-    private Integer courseRestTime;
+    private Long courseRestTime;
 
     /**
      * 上次上课时间
@@ -58,13 +58,13 @@ public class CourseRecord implements Serializable {
      * 课程归属管理群组
      */
     @TableField(value = "course_admin_group_id")
-    private Integer courseAdminGroupId;
+    private Long courseAdminGroupId;
 
     /**
      * 课程归属人
      */
     @TableField(value = "course_owner_user_id")
-    private Integer courseOwnerUserId;
+    private Long courseOwnerUserId;
 
     /**
      * 课程备注
